perplexity+qwen/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
dashscope>=1.10.0
python-dotenv>=0.19.0
streamlit>=1.24.0
httpx>=0.24.0 